requirements.txt
# ********************************************************************************
# Copyright (c) 2025, Wentao Guo, Mayank Mishra, Xinle Cheng, Ion Stoica, Tri Dao
# ********************************************************************************

torch==2.4.0  # Compatible with CUDA 12.4
nvidia-cutlass-dsl==4.4.0
quack-kernels==0.2.5
pytest
parameterized
ninja
rich
